[BluetoothAudioSink] Clean up tracing
